chore: refactor flows
