+++ ui/selectors/qr-sync/qr-sync.ts
@@ -2,7 +2,6 @@ import { createSelector } from 'reselect';
 // eslint-disable-next-line import-x/no-restricted-paths
 import type { QrSyncControllerState } from '../../../app/scripts/controllers/qr-sync/types';
 import {
-  QR_SYNC_PHASES,
   QR_SYNC_TERMINAL_PHASES,
   type QrSyncPhase,
 } from '../../../shared/constants/qr-sync';
